backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
gunicorn==20.1.0  # For production server
Pillow==10.0.0
dj-database-url==1.3.0